refactor: Card and Flip components (#54)
Very rough refactor to be compatible with the upcoming Swipe animation effect.

`<CardFlipTest />` needed to switch from `div` to `animated.div` as for some reason when the Swipe component used `animated(CardFlipTest)` it caused many unwanted re-renders? (but is supposedly meant to be the same?)

`<Flip />` has some added notes to assist with grokking what is going on. It also fixes the vertical flip to horizontal flip correctly :)

`<Card />` has had it's `<Button />` intended for triggering the flip effect extracted out to a `<FlipButton />` component. Plus some additional changes to improve the `<Image />` component used, along with minor layout/style changes (such as the byline gradient fade, and a `<Stack />` to push the button to the bottom of the card, instead of directly after the content before it.